Fixed dependence on file load order
Moved set up of Vec2 and other prototypes inside the Physics constructor so this code is run when calling new Physics() instead of when the file is loaded. This allows for the boxwrap3d.js and Box3D.js files to be imported in any order which is otherwise platform dependent.